+++ DipPlayground.playground/Pages/Resolving components.xcplaygroundpage/Contents.swift
@@ -12,7 +12,7 @@ let container = DependencyContainer { container in
 You resolve previously registered definition using `resolve` method:
 */
 
-var service = container.resolve() as Service
+var service = try! container.resolve() as Service
 
 /*:
 That code says that you want your `container` to give you an instance that was registered as implementation of `Service` protocol.
@@ -20,7 +20,7 @@ That code says that you want your `container` to give you an instance that was r
 It's important to specify the same type that you used for registration. You can use either `as` syntax, or specify type of you variable when you define it:
 */
 
-let otherService: Service = container.resolve()
+let otherService: Service = try! container.resolve()
 
 /*:
 Both ways will let the `container` detect the type that you want to resolve as. We prefer the `as` syntax because it reads more naturally in Swift.
@@ -32,16 +32,16 @@ container.register(tag: "production") { ServiceImp1() as Service }
 container.register(tag: "test") { ServiceImp2() as Service }
 
 // Will give you a ServiceImp1 instance
-let productionService = container.resolve(tag: "production") as Service
+let productionService = try! container.resolve(tag: "production") as Service
 // Will give you a ServiceImp2 instance
-let testService = container.resolve(tag: "test") as Service
+let testService = try! container.resolve(tag: "test") as Service
 // Will give you a ServiceImp1 because one was registered without a tag on line 4
-let defaultService = container.resolve() as Service
+let defaultService = try! container.resolve() as Service
 
 /*:
 You can use runtime arguments to resolve components. Dip supports up to six arguments. For more details see ["Runtime arguments"](Runtime%20arguments).
 */
 container.register { service in ClientImp1(service: service) as Client }
-let client = container.resolve(service) as Client
+let client = try! container.resolve(withArguments: service) as Client
 
 //: [Next: Runtime Arguments](@next)